crud_escolar_api/crud_escolar_api/requirements.txt
Django==5.0.2
django-compat==1.0.15
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5
mysqlclient
pymysql
django-rest-auth
cryptography
python-dateutil
google-cloud-storage